Breakout and re-enable forward param accessors code.
The code got accidentally disabled when refactoring SuperAccessors for
pickling. The forwardParamAccessor method was applied only to non-parameters
where it is the identity (we really shopuld get test paramAccessors working,
then this would not have happened.

In the interest of better modularity, the code was placed in its own trait,
because it overlaps only marginally with the rest of SuperAccessors functionality.
